+++ app/WarningGitHubRateLimiting.tsx
@@ -1,0 +1,13 @@
+'use client';
+
+import React from 'react';
+import { Alert, AlertIcon } from '@chakra-ui/react';
+
+export default function AppWarningGitHubRateLimiting() {
+    return (
+        <Alert status="warning">
+            <AlertIcon />
+            This demo might run into issues caused by GitHub rate-limiting, sorry!
+        </Alert>
+    );
+}
