+++ app/layout.tsx
@@ -4,6 +4,7 @@ import { Metadata } from 'next';
 import { Inter } from 'next/font/google';
 
 import EmotionRootStyleRegistry from './EmotionRootStyleRegistry';
+import AppWarningGitHubRateLimiting from './WarningGitHubRateLimiting';
 import { NavBar } from 'components/NavBar';
 
 // If loading a variable font, you don't need to specify the font weight
@@ -17,10 +18,6 @@ export const metadata: Metadata = {
     icons: '/favicon.ico',
 };
 
-//Later: (as it does not work here with /app)
-//import { ColorModeScript } from '@chakra-ui/react';
-//<ColorModeScript/*initialColorMode={'light' /* | "dark" | "system" * /}*/ />
-
 export default function RootLayout({ children }: { children: React.ReactNode }) {
     return (
         <html lang="en" className={inter.className}>
@@ -29,7 +26,10 @@ export default function RootLayout({ children }: { children: React.ReactNode }) 
                 <EmotionRootStyleRegistry>
                     <>
                         <NavBar />
-                        <div>{children}</div>
+
+                        <AppWarningGitHubRateLimiting />
+
+                        {children}
                     </>
                 </EmotionRootStyleRegistry>
             </body>
