+++ app/restful/layout.tsx
@@ -1,0 +1,6 @@
+import React from 'react';
+import DirectorLayout from 'components/DirectorLayout';
+
+export default function Layout({ children }: { children: React.ReactNode }) {
+    return <DirectorLayout>{children}</DirectorLayout>;
+}
