+++ app/side-by-side/[userName]/[repoName]/side-by-side.module.css
@@ -1,7 +1,6 @@
 .sideBySide {
     display: flex;
     flex-direction: row;
-    _width: 100vw;
 }
 .side {
     width: 50vw;
